Add PIPECE solver

+++ src/FDDE/DelayABM.jl
@@ -20,7 +20,7 @@ struct DelayABM <: FractionalDiffEqAlgorithm end
 function solve(FDDE::FDDEProblem, h, ::DelayABM)
     @unpack f, ϕ, α, τ, T = FDDE
     N::Int = round(Int, T/h)
-    Ndelay = round(Int, τ/h)
+    Ndelay::Int = round(Int, τ/h)
     x1 = zeros(Ndelay+N+1)
     x = zeros(Ndelay+N+1)
     #x1[Ndelay+N+1] = 0
